+++ common/src/animations/double_carrousel.rs
@@ -9,6 +9,7 @@ use rand::{RngExt, SeedableRng};
 use smart_leds::{RGB8, brightness, gamma};
 use smart_leds_trait::SmartLedsWrite;
 
+const BRIGHTNESS_DAMPING_FACTOR: f32 = 0.05;
 const LEDS_PER_CARROUSEL: usize = LEDS_TOTAL / 2;
 
 pub struct DoubleCarrousel {
@@ -43,7 +44,7 @@ impl DoubleCarrousel {
             leds_section_1
                 .write(brightness(
                     gamma(data[LEDS_SECTION_1_RANGE].iter().copied()),
-                    self.brightness(settings),
+                    settings.brightness_damped(BRIGHTNESS_DAMPING_FACTOR),
                 ))
                 .unwrap();
         };
@@ -52,7 +53,7 @@ impl DoubleCarrousel {
             leds_section_2
                 .write(brightness(
                     gamma(data[LEDS_SECTION_2_RANGE].iter().copied()),
-                    self.brightness(settings),
+                    settings.brightness_damped(BRIGHTNESS_DAMPING_FACTOR),
                 ))
                 .unwrap();
         };
@@ -87,9 +88,5 @@ impl DoubleCarrousel {
             }
             self.color_index_2 = new_color;
         }
-    }
-
-    fn brightness(&self, settings: &Settings) -> u8 {
-        (f32::from(settings.brightness()) * 0.05) as u8
     }
 }